[REWORK] Reworked pipe API
Upon switching to the new pipe rendering system, we lost the ability to know a pipe direction, and to separate pipes. This commit aim to fix the problem. Not so long ago, I patched the rendering to take Hue into account. This does the same thing in the pipe api,

CHANGELOG:
- Clean pipes are now orange, so they can be affected by a hue shift, and be visually identifiable.
- Clean pipe are now made of copper, to justify their new color.
- Clean pipe material and item descriptions fixed
- There are now 7 additional recipes for clean pipes, consumming clean pipes, and producing hueShifted pipes.
- pipe API now take hue into account.
- pipes.tileDescription now take a TileDescriptor ({name, hueShift}) as an optional parameter
- pipes.getPipeTileData not rake a TileDescriptor instead of direction as an optional parameter
- some fixes